requirements.txt
Flask==0.11.1
attrdict==2.0.0
eventlet==0.19.0
python-socketio==1.6.1
numpy==1.16.4
Pillow==6.0.0
scipy==0.19.1
keras==2.2.4
tensorflow==1.14.0
h5py==2.6.0

